rag-api-custom/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-openai
langchain-community
langchain-postgres
psycopg2-binary
pydantic_ai